refactor: comment의 jpa 영속화 객체 패키지 변경
