+++ src/test/java/dandi/dandi/notification/adapter/persistence/convertor/NotificationConvertorsTest.java
@@ -7,8 +7,8 @@ import static dandi.dandi.post.PostFixture.POST_ID;
 import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
 import static org.mockito.Mockito.when;
 
-import dandi.dandi.comment.adapter.persistence.CommentJpaEntity;
-import dandi.dandi.comment.adapter.persistence.CommentRepository;
+import dandi.dandi.comment.adapter.out.persistence.jpa.CommentJpaEntity;
+import dandi.dandi.comment.adapter.out.persistence.jpa.CommentRepository;
 import dandi.dandi.notification.adapter.persistence.NotificationJpaEntity;
 import dandi.dandi.notification.domain.Notification;
 import dandi.dandi.notification.domain.NotificationType;
